Include all memory regions in SetVirtualAddressMap
According to the Linux source code some systems have buggy firmware that
accesses memory not marked as runtime regions. Linux has a function
called should_map_region to determine which regions need to be mapped
and which don't need to be. Given that Redox OS mapps the entirety of
the physical memory anyway, let's just tell the firmware about the
virtual address of all memory regions, even those not strictly
needed by the firmware.