Tell the UEFI firmware that everything is identity mapped
According to the Linux source code, some firmware forgets to update some
pointers in SetVirtualAddressMap. Using an identity map fixes this
issue. We still need to call SetVirtualAddressMap anyway as other
firmware misbehaves when it isn't called before using runtime services.

When the Redox kernel starts using runtime services, it will have to
actually fulfill the identity mapping we tell the bootloader. Before
that, it shouldn't matter what the mapping is.